Use `UInt32` for `gitCommitNthGenAncestor()` index parameter
Changed the `n` parameter's type from `UInt` to `UInt32` to remove the
type cast that could potentially overflow on some platforms. Although an
overflow is unlikely in practice, the caller should handle conversions.

+++ Sources/SwiftLibgit2/Commit/Commit-Functions.swift
@@ -543,7 +543,7 @@ public func gitCommitParentID(
 public func gitCommitNthGenAncestor(
     ancestor    : UnsafeMutablePointer<OpaquePointer?>,
     commit      : OpaquePointer,
-    n           : UInt
+    n           : UInt32
 ) -> GitErrorCode
 {
     return withCConversion
@@ -551,7 +551,7 @@ public func gitCommitNthGenAncestor(
         return git_commit_nth_gen_ancestor(
             ancestor,
             commit,
-            UInt32(n)
+            n
         )
     }
 }
